Read whatever config can be from foreman-installer-answers.yaml
The only parameter that the Puppet container does not need to know
about, but the outside plumbing does, is the netmask of the VIP's
network (required by pipework). For now, hard-code its value – The
Right Thing would be to save that data in a separate file, along with
the interactive answers.