Fix broken PS1.
Setting the locale to C.UTF-8 lets bash understand that the "⚠"
character, despite being 3 bytes long, is actually only one character as
far as displaying is concerned. Only if bash is able to
correctly figure out the on-display width of PS1, will it be able to
wrap around lines at the proper time.