Override puppetdb_conn_validator
The version provided with the Puppetdb Foreman module wants to use the
Puppet code flows to perform the SSL probe using the same client
certificates as would an agent. Unfortunately this idea requires that
1) Puppet agent has already run once (which it won't at
foreman-installer time), and 2) that we want to run the agent in the
Dockerized puppetmaster in the first place (which *I* don't).

* At plan time, purge the files that define puppetdb_conn_validator from
  the installed /usr/share/foreman-installer tree
* Provide a define + exec + curl-based replacement from epflsti.pp;
  rely on the bizarre global-include semantics of Puppet so that the
  puppetdb module can actually see this epflsti-provided define, despite
  puppetdb being loaded first.